Remove ClientConnection
