+++ src/Compilers/Server/VBCSCompiler/NamedPipeClientConnection.cs
@@ -62,53 +62,4 @@ namespace Microsoft.CodeAnalysis.CompilerServer
             throw new Exception("Insufficient resources to process new connection.");
         }
     }
-
-    internal sealed class NamedPipeClientConnection : ClientConnection
-    {
-        private readonly NamedPipeServerStream _pipeStream;
-
-        internal NamedPipeClientConnection(ICompilerServerHost compilerServerHost, string loggingIdentifier, NamedPipeServerStream pipeStream)
-            : base(compilerServerHost, loggingIdentifier, pipeStream)
-        {
-            _pipeStream = pipeStream;
-        }
-
-        /// <summary>
-        /// The IsConnected property on named pipes does not detect when the client has disconnected
-        /// if we don't attempt any new I/O after the client disconnects. We start an async I/O here
-        /// which serves to check the pipe for disconnection. 
-        ///
-        /// This will return true if the pipe was disconnected.
-        /// </summary>
-        protected override Task CreateMonitorDisconnectTask(CancellationToken cancellationToken)
-        {
-            return BuildServerConnection.CreateMonitorDisconnectTask(_pipeStream, LoggingIdentifier, cancellationToken);
-        }
-
-        protected override void ValidateBuildRequest(BuildRequest request)
-        {
-            // Now that we've read data from the stream we can validate the identity.
-            if (!NamedPipeUtil.CheckClientElevationMatches(_pipeStream))
-            {
-                throw new Exception("Client identity does not match server identity.");
-            }
-        }
-
-        public override void Close()
-        {
-            CompilerServerLogger.Log($"Pipe {LoggingIdentifier}: Closing.");
-            try
-            {
-                _pipeStream.Close();
-            }
-            catch (Exception e)
-            {
-                // The client connection failing to close isn't fatal to the server process.  It is simply a client
-                // for which we can no longer communicate and that's okay because the Close method indicates we are
-                // done with the client already.
-                var msg = string.Format($"Pipe {LoggingIdentifier}: Error closing pipe.");
-                CompilerServerLogger.LogException(e, msg);
-            }
-        }
-    }
 }
