+++ src/Compilers/Server/VBCSCompilerTests/ClientConnectionTests.cs
@@ -16,7 +16,7 @@ using System.Data.SqlClient;
 
 namespace Microsoft.CodeAnalysis.CompilerServer.UnitTests
 {
-    public class ClientConnectionTests
+    public class NamedPipeClientConnectionTests
     {
         private static readonly BuildRequest s_emptyCSharpBuildRequest = new BuildRequest(
             BuildProtocolConstants.ProtocolVersion,
